test sql group by

+++ Python/reqlist/functions.py
@@ -11,6 +11,10 @@ def restart():
         return
 
     s = "Traitement en cours détecté. Tuer ? (o/n)"
+    if gl.TEST_RESTART:
+        com.log(s)
+        com.log_print("n (test de la fonctionnalité de reprise)")
+        return
     if com.log_input(s) == 'o':
         com.delete_folder(gl.TMP_PATH)
         return
